Improve efficiency of generate-candidates phase.
Instead of removing duplicates and sorting at the end of the process,
switch to a strategy where we sort by distance upon insertion into the
collection, and keep only one candidate of a certain value (the closest
one, if there are multiple) rather than mark dupes after sorting.

After starting with a naïve begin-at-zero insertion strategy, I switched
to a bisection strategy and reduced the number of lookups in the test
suite from around 1000 to around 400. I’m sure I could do better than
that, given that each match tends to be either slightly closer or
slightly farther away than the previous, but that’s good enough for now.

If I did proper benchmarking, I could estimate how much of a performance
improvement this is, but let’s just assume this is an improvement.

Also, we now support the ability to do inline completion when part of a
word is selected; this is also supported in TextMate, and behaves
just like if the selection didn’t exist, with the first suggestion
replacing the original selection.